feat: refactor sections into files, add relationships
