+++ src/components/contentModel/editor/contentModelEditor.component.tsx
@@ -1,17 +1,8 @@
-import clsx from "clsx";
-import { Reorder } from "framer-motion";
-import { ReactNode, useCallback } from "react";
 import { useForm } from "react-hook-form";
-import { FiInfo } from "react-icons/fi";
 import { toast } from "react-toastify";
 
 import { Button } from "src/components/button";
-import { ColourPicker } from "src/components/inputs/colourPicker";
-import { Select } from "src/components/inputs/select";
-import { TextInput } from "src/components/inputs/textInput";
-import { ObjectIdentifierCard } from "src/components/objectIdentifierCard";
 import { Toast } from "src/components/toast/toast.component";
-import { Tooltip } from "src/components/tooltip/tooltip.component";
 import { SYSTEM_FIELDS } from "src/constants/skylark";
 import { useUpdateObjectTypeConfig } from "src/hooks/schema/update/useUpdateObjectTypeConfig";
 import {
@@ -20,24 +11,16 @@ import {
   SkylarkObjectMeta,
   ParsedSkylarkObjectConfigFieldConfig,
   InputFieldWithFieldConfig,
-  SkylarkSystemField,
-  AvailabilityStatus,
 } from "src/interfaces/skylark";
 import { isSkylarkObjectType } from "src/lib/utils";
 
-import { ObjectTypeFieldInput } from "./contentModelRowInput.component";
-
-type FieldSection = "system" | "translatable" | "global";
-
-type FieldSectionObject = Record<
-  "system" | "translatable" | "global",
-  { title: string; fields: InputFieldWithFieldConfig[] }
->;
-
-const sectionClassName = "my-10 border-t pt-10";
-
-const uiDisplayFieldTooltip =
-  "A config property that instructs the UI which field it should use when displaying an object on listing pages.";
+import {
+  ContentModelEditorForm,
+  FieldSectionObject,
+} from "./sections/common.component";
+import { FieldsSection } from "./sections/fields.component";
+import { RelationshipsSection } from "./sections/relationships.component";
+import { UIConfigSection } from "./sections/uiConfig.component";
 
 const combineFieldAndFieldConfigAndSortByConfigPostion = (
   inputFields: NormalizedObjectField[],
@@ -147,111 +130,6 @@ const createFieldSections = (
   return sections;
 };
 
-const InfoTooltip = ({ tooltip }: { tooltip: ReactNode }) => (
-  <Tooltip tooltip={tooltip}>
-    <div className="ml-1">
-      <FiInfo className="text-base" />
-    </div>
-  </Tooltip>
-);
-
-const SectionHeader = ({ children }: { children: ReactNode }) => (
-  <h3 className="text-xl">{children}</h3>
-);
-
-const FieldHeader = ({
-  children,
-  className,
-  tooltip,
-}: {
-  children: ReactNode;
-  className?: string;
-  tooltip?: ReactNode;
-}) => (
-  <div className={clsx("flex items-center whitespace-pre", className)}>
-    <p>{children}</p>
-    {tooltip && <InfoTooltip tooltip={tooltip} />}
-  </div>
-);
-
-const FieldSection = ({
-  title,
-  fields,
-  objectMeta,
-  objectConfig,
-  disableReorder,
-  primaryField,
-  onChange,
-  onPrimaryFieldChange,
-}: {
-  title: string;
-  fields: InputFieldWithFieldConfig[];
-  objectMeta: SkylarkObjectMeta;
-  objectConfig?: ParsedSkylarkObjectConfig;
-  disableReorder?: boolean;
-  primaryField?: string | null;
-  onChange: (fields: InputFieldWithFieldConfig[]) => void;
-  onPrimaryFieldChange: (field: string) => void;
-}) => {
-  const handleChange = useCallback(
-    (updatedFieldWithConfig: InputFieldWithFieldConfig) => {
-      const updatedFields = fields.map((fieldWithConfig) => {
-        if (fieldWithConfig.field.name === updatedFieldWithConfig.field.name) {
-          return {
-            ...fieldWithConfig,
-            ...updatedFieldWithConfig,
-          };
-        }
-        return fieldWithConfig;
-      });
-
-      onChange(updatedFields);
-    },
-    [fields, onChange],
-  );
-
-  return (
-    <div className="mt-4 mb-10">
-      <h4 className="mt-4 mb-2 text-lg font-medium">{title}</h4>
-      <div className="grid grid-cols-7 gap-4 text-manatee-400 font-normal text-sm">
-        <FieldHeader className="col-span-2">Name</FieldHeader>
-        <FieldHeader className="col-span-2" tooltip="The GraphQL type">
-          Type
-        </FieldHeader>
-        <FieldHeader className="col-span-2">Enum / UI type</FieldHeader>
-        <FieldHeader tooltip="When creating an object of this type, this field will be required to be added.">
-          Required
-        </FieldHeader>
-        {/* <FieldHeader tooltip={uiDisplayFieldTooltip}>
-          UI Display field
-        </FieldHeader> */}
-      </div>
-      <Reorder.Group onReorder={onChange} values={fields}>
-        {fields.map((fieldWithConfig) => {
-          const fieldName = fieldWithConfig.field.name;
-          return (
-            <ObjectTypeFieldInput
-              key={`${objectMeta.name}-${fieldName}`}
-              fieldWithConfig={fieldWithConfig}
-              objectMeta={objectMeta}
-              disableReorder={disableReorder}
-              isPrimaryField={
-                primaryField
-                  ? primaryField === fieldName
-                  : objectConfig?.primaryField === fieldName
-              }
-              onChange={handleChange}
-              onPrimaryFieldCheckedChange={(checked) =>
-                onPrimaryFieldChange(checked ? fieldName : "")
-              }
-            />
-          );
-        })}
-      </Reorder.Group>
-    </div>
-  );
-};
-
 export const ObjectTypeEditor = ({
   objectMeta,
   objectConfig,
@@ -259,12 +137,7 @@ export const ObjectTypeEditor = ({
   objectMeta: SkylarkObjectMeta;
   objectConfig?: ParsedSkylarkObjectConfig;
 }) => {
-  const form = useForm<{
-    fieldSections: FieldSectionObject;
-    objectTypeDisplayName: string;
-    primaryField: string | undefined | null;
-    colour: string | undefined | null;
-  }>({
+  const form = useForm<ContentModelEditorForm>({
     // Can't use onSubmit because we don't have a submit button within the form
     mode: "onTouched",
     values: {
@@ -333,19 +206,6 @@ export const ObjectTypeEditor = ({
     )();
   };
 
-  // TODO improve by passing form into components so that we don't have to use a watch this high
-  const { fieldSections, objectTypeDisplayName, primaryField, colour } =
-    form.watch();
-
-  const onFieldChange = useCallback(
-    (id: FieldSection, reorderedFields: InputFieldWithFieldConfig[]) => {
-      form.setValue(`fieldSections.${id}.fields`, reorderedFields, {
-        shouldDirty: true,
-      });
-    },
-    [form],
-  );
-
   return (
     <div key={objectMeta.name} className="">
       <div className="flex justify-between mb-10">
@@ -376,126 +236,13 @@ export const ObjectTypeEditor = ({
           </Button>
         </div>
       </div>
-      <section className={sectionClassName}>
-        <SectionHeader>UI Config</SectionHeader>
-        <div className="flex flex-col md:grid md:space-x-8 max-md:space-y-8 grid-cols-2 text-sm text-manatee-600">
-          <div className="grid grid-cols-3 w-full items-center auto-rows-fr gap-x-2 mt-4 gap-y-1">
-            <FieldHeader
-              className="col-span-1"
-              tooltip="A mask for the Object Type, in case you want it to display differently in the UI. Useful when you have Object Types with many characters."
-            >
-              Display name
-            </FieldHeader>
-            {/* <p className="text-black">{objectTypeDisplayName}</p> */}
-            <div className="col-span-2">
-              <TextInput
-                value={objectTypeDisplayName}
-                onChange={(str) =>
-                  form.setValue("objectTypeDisplayName", str, {
-                    shouldDirty: true,
-                  })
-                }
-              />
-            </div>
-            <FieldHeader className="col-span-1" tooltip={uiDisplayFieldTooltip}>
-              Display field
-            </FieldHeader>
-            <Select
-              className="col-span-2"
-              variant="primary"
-              options={[
-                ...fieldSections.system.fields,
-                ...fieldSections.translatable.fields,
-                ...fieldSections.global.fields,
-              ].map(({ field }) => ({ label: field.name, value: field.name }))}
-              placeholder=""
-              selected={primaryField || SkylarkSystemField.UID}
-              onChange={(value) =>
-                form.setValue("primaryField", value, {
-                  shouldDirty: true,
-                })
-              }
-            />
-            <FieldHeader
-              className="col-span-1"
-              tooltip="Colour to represent the Object Type in the UI. Useful to identify an Object Type at a glance."
-            >
-              Colour
-            </FieldHeader>
-            <div className="col-span-2">
-              <ColourPicker
-                colour={colour || ""}
-                onChange={(colour) =>
-                  form.setValue("colour", colour, { shouldDirty: true })
-                }
-              >
-                <span className="group-hover/colour-picker:text-brand-primary text-xs group-hover/colour-picker:underline text-manatee-400">
-                  Change
-                </span>
-              </ColourPicker>
-            </div>
-            <FieldHeader
-              className="col-span-1"
-              tooltip="Controls the display order of fields in the UI. System fields cannot be reordered."
-            >
-              Order
-            </FieldHeader>
-            <p className="col-span-2">Drag the fields below to reorder.</p>
-          </div>
-          <div className="flex justify-center items-center">
-            <div className="flex justify-center flex-col w-full pr-4">
-              <p className="mb-2">Resulting Object Type overview:</p>
-              <ObjectIdentifierCard
-                forceConfigFromObject
-                className="shadow px-2"
-                object={{
-                  objectType: objectMeta.name,
-                  uid: "example",
-                  config: {
-                    primaryField: null,
-                    colour,
-                    objectTypeDisplayName,
-                  },
-                  meta: {
-                    language: "",
-                    availableLanguages: [],
-                    availabilityStatus: AvailabilityStatus.Active,
-                  },
-                  metadata: {
-                    uid: `Example "${primaryField}" value`,
-                    external_id: "",
-                  },
-                  availability: {
-                    status: AvailabilityStatus.Active,
-                    objects: [],
-                  },
-                }}
-              />
-            </div>
-          </div>
-        </div>
-      </section>
-      <section className={sectionClassName}>
-        <SectionHeader>Fields</SectionHeader>
-        {/* TODO add Skeleton here with hardcoded headings as sections headers can be filtered out when no fields exist */}
-        {Object.entries(fieldSections).map(([id, { title, fields }]) => (
-          <FieldSection
-            key={id}
-            title={title}
-            fields={fields}
-            objectMeta={objectMeta}
-            objectConfig={objectConfig}
-            disableReorder={id === "system"}
-            primaryField={primaryField}
-            onChange={(fieldsWithConfig) =>
-              onFieldChange(id as FieldSection, fieldsWithConfig)
-            }
-            onPrimaryFieldChange={(field) =>
-              form.setValue("primaryField", field, { shouldDirty: true })
-            }
-          />
-        ))}
-      </section>
+      <UIConfigSection form={form} objectMeta={objectMeta} />
+      <FieldsSection
+        form={form}
+        objectMeta={objectMeta}
+        objectConfig={objectConfig}
+      />
+      <RelationshipsSection form={form} objectMeta={objectMeta} />
     </div>
   );
 };
