+++ src/components/pill/objectTypePill/objectTypePill.component.tsx
@@ -9,12 +9,14 @@ interface ObjectTypePillProps {
   type: SkylarkObjectType;
   defaultConfig?: ParsedSkylarkObjectConfig;
   className?: string;
+  forceActualName?: boolean;
 }
 
 export const ObjectTypePill = ({
   type,
   defaultConfig,
   className,
+  forceActualName,
 }: ObjectTypePillProps) => {
   const { objectTypesWithConfig } = useSkylarkObjectTypesWithConfig();
 
@@ -24,7 +26,7 @@ export const ObjectTypePill = ({
 
   return (
     <Pill
-      label={config?.objectTypeDisplayName || type}
+      label={forceActualName ? type : config?.objectTypeDisplayName || type}
       bgColor={config?.colour || undefined}
       className={className}
     />
